Fixed DELETE /api/event/tag (release)

+++ backend/src/controllers/eventController.js
@@ -158,23 +158,31 @@ const deleteEventTag = async (req, res) => {
 	if (!isOwner)
 		return res.status(400).json({ message: "You dont own this event tag!" });
 
-	if (isDefault)
-		return res
-			.status(400)
-			.json({ message: "You can't delete the default tag" });
-
-	// Remove tag from user
-	user.eventTags.splice(user.eventTags.indexOf(tag), 1);
-
 	// Remove events from user
 	const events = tag.appliedTo;
 	events.forEach((e) => user.events.splice(user.events.indexOf(e), 1));
 
+	if (!isDefault)
+		// Remove tag from user
+		user.eventTags.splice(user.eventTags.indexOf(tag), 1);
+	else tag.appliedTo = [];
+
 	try {
+		// Remove events
 		for (const e of events) {
 			await Event.findByIdAndDelete(e);
+			// Remove events from users that the event is shared with
+			for (const id of e.sharedWith) {
+				const u = await User.findById(id);
+				u.sharedEvents.splice(u.sharedEvents.indexOf(e), 1);
+				await u.save();
+			}
 		}
-		await EventTag.findByIdAndDelete(tag);
+
+		// Remove tag
+		if (!isDefault) await EventTag.findByIdAndDelete(tag);
+		else await tag.save();
+
 		await user.save();
 
 		return res.json({ message: "success" });
